Restructure kernel-ewasm to seperate crate + setup workspace
