+++ kernel-ewasm/Cargo.toml
@@ -1,32 +1,9 @@
-[package]
-description = "Cap9 Ewasm Kernel Library"
-name = "kernel-ewasm"
-version = "0.1.0"
-authors = ["Daohub Inc <info@daohub.io>"]
-edition = "2018"
+[workspace]
 
-[dependencies]
-tiny-keccak = "1.4.2"
-pwasm-std = "0.13"
-pwasm-ethereum = "0.8"
-pwasm-abi = "0.2"
-pwasm-abi-derive = "0.2"
-lazy_static = { version = "1.2.0", features = ["spin_no_std"] }
-cap9-std = { path = "cap9-std" }
-
-[dev-dependencies.pwasm-test]
-git = "https://github.com/paritytech/pwasm-test"
-default-features = false
-
-[lib]
-name = "kernel_ewasm"
-crate-type = ["cdylib"]
-
-[[examples]]
-crate-type = ["cdylib"]
-
-[features]
-std = ["pwasm-std/std", "pwasm-ethereum/std", "pwasm-test/std"]
+members = [
+    "cap9-kernel",
+    "cap9-std"
+]
 
 [profile.release]
 panic = "abort"
