+++ kernel-ewasm/build.bat
@@ -1,7 +1,9 @@
 
 cargo build --release --target wasm32-unknown-unknown
-wasm-build --target=wasm32-unknown-unknown .\target kernel-ewasm
+cargo build --examples --target wasm32-unknown-unknown --release --features std
 
-mkdir .\build
-copy .\target\*.wasm .\build
-copy .\target\json\* .\build
+COPY ./target/wasm32-unknown-unknown/release/examples/*_test.wasm ./target/wasm32-unknown-unknown/release
+wasm-build --target=wasm32-unknown-unknown ./target cap9-kernel
+
+wasm-build --target=wasm32-unknown-unknown ./target writer_test
+wasm-build --target=wasm32-unknown-unknown ./target entry_test
