+++ kernel-ewasm/cap9-kernel/src/proc_table/mod.rs
@@ -643,7 +643,6 @@ pub mod contract {
 #[allow(non_snake_case)]
 mod tests {
     extern crate pwasm_test;
-    extern crate std;
 
     use super::cap::*;
     use super::contract;
