+++ kernel-ewasm/cap9-std/src/lib.rs
@@ -1,3 +1,5 @@
+#![no_std]
+
 /// Generic wasm error
 #[derive(Debug)]
 pub struct Error;
